+++ src/main/java/online/kingdomkeys/kingdomkeys/client/gui/menu/items/equipment/MenuArmorSelectorScreen.java
@@ -87,8 +87,8 @@ public class MenuArmorSelectorScreen extends MenuBackground {
 				}
 			}
 		}
-		keyblades = new MenuBox((int) keybladesX, (int) keybladesY, (int) keybladesWidth, (int) keybladesHeight, colour);
-		details = new MenuBox((int) detailsX, (int) keybladesY, (int) detailsWidth, (int) keybladesHeight, colour);
+		keyblades = new MenuBox((int) keybladesX, (int) keybladesY, (int) keybladesWidth, (int) keybladesHeight,0.6F, colour);
+		details = new MenuBox((int) detailsX, (int) keybladesY, (int) detailsWidth, (int) keybladesHeight,0.6F, colour);
 	}
 
 	@Override