+++ src/main/java/online/kingdomkeys/kingdomkeys/client/gui/menu/items/equipment/MenuShotlockSelectorScreen.java
@@ -72,7 +72,7 @@ public class MenuShotlockSelectorScreen extends MenuBackground {
 			}
 		}
 
-		keyblades = new MenuBox((int) keybladesX, (int) keybladesY, (int) keybladesWidth, (int) keybladesHeight, colour);
+		keyblades = new MenuBox((int) keybladesX, (int) keybladesY, (int) keybladesWidth, (int) keybladesHeight, 0.6F,colour);
 		//details = new MenuBox((int) detailsX, (int) keybladesY, (int) detailsWidth, (int) keybladesHeight, colour);
 	}