+++ src/main/java/online/kingdomkeys/kingdomkeys/client/gui/synthesis/SynthesisForgeScreen.java
@@ -23,6 +23,7 @@ import online.kingdomkeys.kingdomkeys.client.sound.ModSounds;
 import online.kingdomkeys.kingdomkeys.data.PlayerData;
 import online.kingdomkeys.kingdomkeys.item.KeybladeItem;
 import online.kingdomkeys.kingdomkeys.item.KeychainItem;
+import online.kingdomkeys.kingdomkeys.item.ModItems;
 import online.kingdomkeys.kingdomkeys.lib.Strings;
 import online.kingdomkeys.kingdomkeys.network.PacketHandler;
 import online.kingdomkeys.kingdomkeys.network.cts.CSCloseMoogleGUI;
@@ -44,7 +45,6 @@ public class SynthesisForgeScreen extends MenuFilterable {
 	int itemsX = 100, itemsY = 100, itemWidth = 140, itemHeight = 10;
 
 	Button upgrade;
-	int itemsPerPage = 10;
 	private MenuButton back;
 	SynthesisScreen parent;
 
@@ -117,18 +117,17 @@ public class SynthesisForgeScreen extends MenuFilterable {
 		float topBarHeight = (float) height * 0.17F;
 		float boxWidth = (float) width * 0.3F;
 		float middleHeight = (float) height * 0.6F;
-		boxL = new MenuBox((int) boxPosX, (int) topBarHeight, (int) boxWidth, (int) middleHeight, new Color(4, 4, 68));
-		boxM = new MenuBox((int) boxPosX + (int) boxWidth, (int) topBarHeight, (int) (boxWidth*0.7F), (int) middleHeight, new Color(4, 4, 68));
-		boxR = new MenuBox(boxM.getX() + (int) (boxWidth*0.7F), (int) topBarHeight, (int) (boxWidth*1.17F), (int) middleHeight, new Color(4, 4, 68));
+		boxL = new MenuBox((int) boxPosX, (int) topBarHeight, (int) boxWidth, (int) middleHeight, 1, new Color(40, 4, 255));
+		boxM = new MenuBox((int) boxPosX + (int) boxWidth, (int) topBarHeight, (int) (boxWidth*0.7F), (int) middleHeight,1F, new Color(108, 40, 40));
+		boxR = new MenuBox(boxM.getX() + (int) (boxWidth*0.7F), (int) topBarHeight, (int) (boxWidth*1.17F), (int) middleHeight,1, new Color(4, 68, 4));
 		int scrollTop = (int) topBarHeight;
 		int scrollBot = (int) (scrollTop + middleHeight);
 
-		//float filterPosX = width * 0.3F;
-		//float filterPosY = height * 0.02F;
-		//filterBar = new MenuFilterBar((int) filterPosX, (int) filterPosY, this);
-		//filterBar.init();
 		scrollBar = new MenuScrollBar((int) (boxPosX + boxWidth - 17), scrollTop, scrollBot, (int) middleHeight, 0);
 		addRenderableWidget(scrollBar);
+		scrollBar2 = new MenuScrollBar(boxR.getX()+boxR.getWidth()- 17, scrollTop, scrollBot, (int) middleHeight, 0);
+		addRenderableWidget(scrollBar2);
+
 
 		initItems();
 
@@ -137,7 +136,6 @@ public class SynthesisForgeScreen extends MenuFilterable {
 		addRenderableWidget(back = new MenuButton((int)this.buttonPosX, this.buttonPosY, (int)buttonWidth, Component.translatable(Strings.Gui_Menu_Back).getString(), MenuButton.ButtonType.BUTTON, b -> minecraft.setScreen(new SynthesisScreen(parent.playerData, parent.invFile, parent.name, parent.moogle))));
 
 		super.init();
-		itemsPerPage = (int) (middleHeight / 14);
 	}
 
 	@Override
@@ -161,12 +159,16 @@ public class SynthesisForgeScreen extends MenuFilterable {
 		items.sort(Comparator.comparing(Utils::getCategoryForStack).thenComparing(stack -> stack.getHoverName().getContents().toString()));
 
 		for (int i = 0; i < items.size(); i++) {
+			MenuStockItem item;
 			if(items.get(i).getItem() instanceof KeychainItem) {
 				KeybladeItem kb = ((KeychainItem)items.get(i).getItem()).toSummon();
-				inventory.add(new MenuStockItem(this, items.get(i), (int) invPosX, (int) invPosY + (i * 14),boxL.getWidth()-scrollBar.getWidth()-6, false, new ItemStack(kb).getHoverName().getString()));
+				item = new MenuStockItem(this, items.get(i), (int) invPosX, (int) invPosY + (i * 14),boxL.getWidth()-scrollBar.getWidth()-6, false, new ItemStack(kb).getHoverName().getString());
 			} else {
-				inventory.add(new MenuStockItem(this, items.get(i), (int) invPosX, (int) invPosY + (i * 14),boxL.getWidth()-scrollBar.getWidth()-6, false));
+				item = new MenuStockItem(this, items.get(i), (int) invPosX, (int) invPosY + (i * 14),boxL.getWidth()-scrollBar.getWidth()-6, false);
 			}
+			item.setBackgroundColor(new Color(30,30,100));
+			inventory.add(item);
+
 		}
 		
 		inventory.forEach(this::addWidget);
@@ -177,7 +179,7 @@ public class SynthesisForgeScreen extends MenuFilterable {
 
         addRenderableWidget(upgrade = Button.builder(Component.translatable(Utils.translateToLocal(Strings.Gui_Synthesis_Forge_Upgrade)), (e) -> {
 			action("upgrade");
-		}).bounds(boxM.getX()+3, (int) (height * 0.67), 70, 20).build());
+		}).bounds(boxM.getX()+boxM.getWidth()/2-35, (int) (height * 0.67), 70, 20).build());
 	}
 
 	@Override
@@ -297,20 +299,28 @@ public class SynthesisForgeScreen extends MenuFilterable {
 			//Materials display
 			matrixStack.pushPose();
 			{
-				matrixStack.translate(iconPosX + 20, height*0.2, 1);
 				if(kb.getKeybladeLevel(selectedItemStack) < kb.getMaxLevel()) {
-					Iterator<Entry<Item, Integer>> itMats = kb.data.getLevelData(kb.getKeybladeLevel(selectedItemStack)).getMaterialList().entrySet().iterator();
+					Iterator<Entry<Item, Integer>> materials = kb.data.getLevelData(kb.getKeybladeLevel(selectedItemStack)).getMaterialList().entrySet().iterator();
 					int i = 0;
-					while(itMats.hasNext()) {
-						Entry<Item, Integer> m = itMats.next();
+					int listHeight = (boxR.getPosY()+10 + (kb.data.getLevelData(kb.getKeybladeLevel(selectedItemStack)).getMaterialList().size()+1)*16)+10 - boxR.getPosY()+10;
+					scrollBar2.setContentHeight(listHeight);
+
+					int startX = (int) iconPosX + 20;
+					int startY = boxR.getPosY()+10;
+					gui.enableScissor(boxR.getX()+2,scrollBar2.getY()+2,boxR.getX()+boxR.getWidth(),scrollBar2.getHeight()-5); //Arbitrary number to hide the cut one
+					while(materials.hasNext()) {
+						Entry<Item, Integer> m = materials.next();
 						ItemStack stack = new ItemStack(m.getKey());
 						String n = Utils.translateToLocal(stack.getDescriptionId());
-						//playerData.setMaterial(m.getKey(), 1);
+						String mats = " x"+m.getValue()+" ("+playerData.getMaterialAmount(m.getKey())+")";
+
 						int color = playerData.getMaterialAmount(m.getKey()) >= m.getValue() ?  0x00FF00 : 0xFF0000;
-						gui.drawString(minecraft.font, n+" x"+m.getValue()+" ("+playerData.getMaterialAmount(m.getKey())+")", 0, (i*16), color);
-						ClientUtils.drawItemAsIcon(stack, matrixStack, -17, (i*16)-4, 16);
+						ClientUtils.drawScrollingText(gui,minecraft.font,Component.literal(n), startX, scrollBar2.getX() - minecraft.font.width(mats), startY + (int) ((i*16)-scrollBar2.scrollOffset), color);
+						gui.drawString(minecraft.font, mats, scrollBar2.getX() - minecraft.font.width(mats), startY + (int) ((i*16)-scrollBar2.scrollOffset), color);
+						ClientUtils.drawItemAsIcon(stack, matrixStack, startX -17, boxR.getPosY()+10 + (int)((i*16)-4-scrollBar2.scrollOffset), 16);
 						i++;
 					}
+					gui.disableScissor();
 				}
 			}
 			matrixStack.popPose();
@@ -333,6 +343,8 @@ public class SynthesisForgeScreen extends MenuFilterable {
 	@Override
 	public boolean mouseClicked(double mouseX, double mouseY, int mouseButton) {
 		scrollBar.mouseClicked(mouseX, mouseY, mouseButton);
+		scrollBar2.mouseClicked(mouseX, mouseY, mouseButton);
+
 		if (mouseButton == 1) {
 			PacketHandler.sendToServer(new CSOpenMenu());
 		}
@@ -342,12 +354,16 @@ public class SynthesisForgeScreen extends MenuFilterable {
 	@Override
 	public boolean mouseReleased(double pMouseX, double pMouseY, int pButton) {
 		scrollBar.mouseReleased(pMouseX, pMouseY, pButton);
+		scrollBar2.mouseReleased(pMouseX, pMouseY, pButton);
+
 		return super.mouseReleased(pMouseX, pMouseY, pButton);
 	}
 
 	@Override
 	public boolean mouseDragged(double pMouseX, double pMouseY, int pButton, double pDragX, double pDragY) {
 		scrollBar.mouseDragged(pMouseX, pMouseY, pButton, pDragX, pDragY);
+		scrollBar2.mouseDragged(pMouseX, pMouseY, pButton, pDragX, pDragY);
+
 		updateScroll();
 		return super.mouseDragged(pMouseX, pMouseY, pButton, pDragX, pDragY);
 	}
@@ -360,7 +376,11 @@ public class SynthesisForgeScreen extends MenuFilterable {
 
 	@Override
 	public boolean mouseScrolled(double mouseX, double mouseY, double deltaX, double deltaY) {
-		scrollBar.mouseScrolled(mouseX, mouseY, deltaX, deltaY);
+		if(mouseX >= boxL.getX() && mouseX <= scrollBar.getX()+ scrollBar.getWidth())
+			scrollBar.mouseScrolled(mouseX, mouseY, deltaX, deltaY);
+		if(mouseX >= boxR.getX() && mouseX <= scrollBar2.getX()+ scrollBar2.getWidth())
+			scrollBar2.mouseScrolled(mouseX, mouseY, deltaX, deltaY);
+
 		updateScroll();
 		return false;
 	}