+++ src/main/java/online/kingdomkeys/kingdomkeys/client/gui/synthesis/SynthesisMaterialScreen.java
@@ -5,6 +5,7 @@ import com.mojang.blaze3d.vertex.PoseStack;
 import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.client.gui.components.Button;
 import net.minecraft.client.gui.components.EditBox;
+import net.minecraft.client.gui.components.Renderable;
 import net.minecraft.client.player.LocalPlayer;
 import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.network.chat.Component;
@@ -18,6 +19,7 @@ import online.kingdomkeys.kingdomkeys.client.gui.elements.MenuBox;
 import online.kingdomkeys.kingdomkeys.client.gui.elements.MenuFilterable;
 import online.kingdomkeys.kingdomkeys.client.gui.elements.buttons.MenuButton;
 import online.kingdomkeys.kingdomkeys.client.gui.elements.buttons.MenuButton.ButtonType;
+import online.kingdomkeys.kingdomkeys.client.gui.elements.buttons.MenuScrollBar;
 import online.kingdomkeys.kingdomkeys.client.gui.elements.buttons.MenuStockItem;
 import online.kingdomkeys.kingdomkeys.client.sound.ModSounds;
 import online.kingdomkeys.kingdomkeys.data.PlayerData;
@@ -26,6 +28,7 @@ import online.kingdomkeys.kingdomkeys.lib.Tags;
 import online.kingdomkeys.kingdomkeys.network.PacketHandler;
 import online.kingdomkeys.kingdomkeys.network.cts.CSCloseMoogleGUI;
 import online.kingdomkeys.kingdomkeys.network.cts.CSDepositMaterials;
+import online.kingdomkeys.kingdomkeys.network.cts.CSOpenMenu;
 import online.kingdomkeys.kingdomkeys.network.cts.CSTakeMaterials;
 import online.kingdomkeys.kingdomkeys.util.Utils;
 import org.jetbrains.annotations.NotNull;
@@ -40,13 +43,12 @@ import java.util.Map.Entry;
 public class SynthesisMaterialScreen extends MenuFilterable {
 		
 	MenuButton deposit,back;
-	Button prev,next, take;
+	Button take;
 	MenuBox boxL, boxR;
 	EditBox amountBox;
 	
 	int page = 0;
-	int itemsPerPage = 14;
-		
+
 	SynthesisScreen parent;
 
 	public SynthesisMaterialScreen(PlayerData playerData, SynthesisScreen parent) {
@@ -65,19 +67,6 @@ public class SynthesisMaterialScreen extends MenuFilterable {
     	super.action(stackRL, stack);
 		int amount = PlayerData.get(minecraft.player).getMaterialAmount(stack.getItem());
 		amountBox.setValue(""+Math.min(64, amount));
-	}
-
-	@Override
-	public boolean mouseScrolled(double mouseX, double mouseY, double deltaX, double deltaY) {
-		if (deltaY > 0 && prev.visible) {
-			action("prev");
-			return true;
-		} else if (deltaY < 0 && next.visible) {
-			action("next");
-			return true;
-		}
-
-		return false;
 	}
 
 	protected void action(String string) {
@@ -127,27 +116,22 @@ public class SynthesisMaterialScreen extends MenuFilterable {
 			break;
 		}
 		
-		updateButtons();
 	}
 
-	private void updateButtons() {
-
-	}
-
-	
 	@Override
 	public void init() {
 		float boxPosX = (float) width * 0.2F;
 		float topBarHeight = (float) height * 0.17F;
 		float boxWidth = (float) width * 0.33F;
 		float middleHeight = (float) height * 0.6F;
-		boxL = new MenuBox((int) boxPosX, (int) topBarHeight, (int) boxWidth, (int) middleHeight, new Color(4, 4, 68));
-		boxR = new MenuBox(boxL.getX() + boxL.getWidth(), (int) topBarHeight, (int) (boxWidth), (int) middleHeight, new Color(4, 4, 68));
-		
+		boxL = new MenuBox((int) boxPosX, (int) topBarHeight, (int) boxWidth, (int) middleHeight,0.6F, new Color(4, 4, 68));
+		boxR = new MenuBox(boxL.getX() + boxL.getWidth(), (int) topBarHeight, (int) (boxWidth), (int) middleHeight, 0.6F,new Color(4, 4, 68));
+
+		scrollBar = new MenuScrollBar((int) (boxPosX + boxWidth - 17), boxL.getY(), boxL.getPosY()+boxL.getHeight(), (int) middleHeight, 0);
+		addRenderableWidget(scrollBar);
+
 		super.init();
 		initItems();
-
-		itemsPerPage = (int) (middleHeight / 14);
 	}
 	
 
@@ -174,21 +158,12 @@ public class SynthesisMaterialScreen extends MenuFilterable {
 		items.sort(Comparator.comparing(Utils::getCategoryForStack).thenComparing(ItemStack::getDescriptionId));
 
 		for (int i = 0; i < items.size(); i++) {
-			inventory.add(new MenuStockItem(this, items.get(i), (int) invPosX, (int) invPosY + (i * 14), (int)(width * 0.3255F), true));
+			MenuStockItem item = new MenuStockItem(this, items.get(i), (int) invPosX, (int) invPosY + (i * 14), boxL.getWidth()-scrollBar.getWidth()-4, true);
+			item.setBackgroundColor(new Color(30,30,100));
+			inventory.add(item);
 		}
-		
 		inventory.forEach(this::addWidget);
 
-		addRenderableWidget(prev = Button.builder(Component.literal("<--"), (e) -> {
-			action("prev");
-		}).bounds((int) buttonPosX + 10, (int)(height * 0.1F), 30, 20).build());
-
-		addRenderableWidget(next = Button.builder(Component.literal("-->"), (e) -> {
-			action("next");
-		}).bounds((int) buttonPosX + 10 + 76, (int)(height * 0.1F), 30, 20).build());
-		
-		prev.visible = false;
-		next.visible = false;
 		addRenderableWidget(deposit = new MenuButton((int) buttonPosX, button_statsY, (int) buttonWidth, Utils.translateToLocal(Strings.Gui_Synthesis_Materials_Deposit), ButtonType.BUTTON, (e) -> { action("deposit"); }));
 		addRenderableWidget(back = new MenuButton((int) buttonPosX, button_statsY + (18), (int) buttonWidth, Utils.translateToLocal(Strings.Gui_Menu_Back), ButtonType.BUTTON, (e) -> { action("back"); }));
 		addRenderableWidget(amountBox = new EditBox(minecraft.font, boxR.getX()+50, (int) (topBarHeight + middleHeight - 20), minecraft.font.width("#####"), 16, Component.translatable("test")) {
@@ -210,7 +185,6 @@ public class SynthesisMaterialScreen extends MenuFilterable {
 		}).bounds(amountBox.getX() + amountBox.getWidth()+5, (int) (topBarHeight + middleHeight - 22), 80, 20).build());
 		
 		take.visible = false;
-		updateButtons();
 	}
 
 	
@@ -224,9 +198,10 @@ public class SynthesisMaterialScreen extends MenuFilterable {
 		gui.setColor(1, 1, 1, 1);
 		super.render(gui, mouseX, mouseY, partialTicks);
 
-		prev.visible = page > 0;
-		next.visible = page < inventory.size() / itemsPerPage;
-		
+		int listHeight = (inventory.get(inventory.size()-1).getY()+20) - inventory.get(0).getY() + 3;
+		scrollBar.setContentHeight(listHeight);
+
+
 		if(minecraft.player.getInventory().getFreeSlot() == -1) { //TODO somehow make this detect in singleplayer the inventory changes
 			take.active = false;
 			take.setMessage(Component.translatable(Strings.Gui_Shop_NoSpace));
@@ -239,24 +214,22 @@ public class SynthesisMaterialScreen extends MenuFilterable {
 			gui.drawString(minecraft.font, Utils.translateToLocal(Strings.Gui_Shop_Page)+" " + (page + 1), 0, 10, 0xFF9900);
 		}
 		matrixStack.popPose();
-		
-		for (int i = 0; i < inventory.size(); i++) {
-			inventory.get(i).active = false;
-		}
 
-		for (int i = page * itemsPerPage; i < page * itemsPerPage + itemsPerPage; i++) {
-			if (i < inventory.size() && i >= 0) {
-				if (inventory.get(i) != null) {
-					inventory.get(i).visible = true;
-					inventory.get(i).setY((int) (topBarHeight) + (i % itemsPerPage) * 14 + 5); // 6 = offset
-					inventory.get(i).render(gui, mouseX, mouseY, partialTicks);
-					inventory.get(i).active = true;
-				}
+        for (MenuStockItem stockItem : inventory) {
+            stockItem.active = false;
+        }
+
+		for(Renderable renderable : this.inventory){
+			if(renderable instanceof MenuStockItem menuStockItem){
+				menuStockItem.active = true;
+				gui.enableScissor(boxL.getX()+2,scrollBar.getY()+2,boxL.getX()+boxL.getWidth(),scrollBar.getHeight()-5); //Arbitrary number to hide the cut one
+				renderable.render(gui,mouseX,mouseY,partialTicks);
+				gui.disableScissor();
+			} else {
+				renderable.render(gui,mouseX,mouseY,partialTicks);
 			}
 		}
-		
-		prev.render(gui, mouseX,  mouseY,  partialTicks);
-		next.render(gui, mouseX,  mouseY,  partialTicks);
+
 		deposit.render(gui, mouseX,  mouseY,  partialTicks);
 		back.render(gui, mouseX,  mouseY,  partialTicks);
 	}
@@ -266,9 +239,7 @@ public class SynthesisMaterialScreen extends MenuFilterable {
 		PoseStack matrixStack = gui.pose();
 		amountBox.render(gui, mouseX,  mouseY,  partialTicks);
 		take.render(gui, mouseX, mouseY, partialTicks);
-		
-		//amountBox.setWidth(minecraft.fontRenderer.getStringWidth(amountBox.getText()));
-		
+
 		take.visible = true;
 
 		float iconPosX = boxR.getX();
@@ -293,7 +264,46 @@ public class SynthesisMaterialScreen extends MenuFilterable {
 		matrixStack.popPose();
 
 	}
-	
+	@Override
+	public boolean mouseClicked(double mouseX, double mouseY, int mouseButton) {
+		scrollBar.mouseClicked(mouseX, mouseY, mouseButton);
+
+		if (mouseButton == 1) {
+			PacketHandler.sendToServer(new CSOpenMenu());
+		}
+		return super.mouseClicked(mouseX, mouseY, mouseButton);
+	}
+
+	@Override
+	public boolean mouseReleased(double pMouseX, double pMouseY, int pButton) {
+		scrollBar.mouseReleased(pMouseX, pMouseY, pButton);
+
+		return super.mouseReleased(pMouseX, pMouseY, pButton);
+	}
+
+	@Override
+	public boolean mouseDragged(double pMouseX, double pMouseY, int pButton, double pDragX, double pDragY) {
+		scrollBar.mouseDragged(pMouseX, pMouseY, pButton, pDragX, pDragY);
+
+		updateScroll();
+		return super.mouseDragged(pMouseX, pMouseY, pButton, pDragX, pDragY);
+	}
+
+	public void updateScroll() {
+		inventory.forEach(button -> {
+			button.offsetY = (int) scrollBar.scrollOffset;
+		});
+	}
+
+	@Override
+	public boolean mouseScrolled(double mouseX, double mouseY, double deltaX, double deltaY) {
+		if(mouseX >= boxL.getX() && mouseX <= scrollBar.getX()+ scrollBar.getWidth())
+			scrollBar.mouseScrolled(mouseX, mouseY, deltaX, deltaY);
+
+		updateScroll();
+		return false;
+	}
+
 	@Override
 	public boolean isPauseScreen() {
 		return false;