+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 **Added:**
 * Experience boost will now boost the amount of vanilla EXP obtained when HP is 50% or below.
 * Going into a Drive Form will refill the MP indicator as well.
-* Revamped Synthesis GUI.
+* Revamped Synthesis, Keyblade Forge and Materials GUIs.
+
+**Changed:**
+* Kiblade recipe to be as long as it used to be a few years ago.
 
 **Fixed:**
 * Airstep not working.
@@ -9,3 +12,4 @@
 * Single shotlock selector and airstep icon not rendering properly.
 * Synthesis bags will no longer be clickable from a synthesis bag inventory screen. 
 * Magic Command Menu being too dark.
+* Shortcuts not being displayed on top left.